requirements.txt
streamlit
langchain
langchain-community
langchain-groq 
PyMuPDF
pytesseract
Pillow
faiss-cpu
sentence-transformers